# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Invoice Capture MVP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /imports:
    get:
      tags: [imports]
      operationId: listImports
      summary: List import batch history
      parameters:
        - name: importType
          in: query
          schema:
            type: string
            enum: [VENDOR_MASTER, PO_REFERENCE, INVOICE_CORRECTION]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Import batches
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportBatchListResponse"
    post:
      tags: [imports]
      operationId: commitImport
      summary: Commit a validated import (applies rows to the database)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportCommitInput"
      responses:
        "201":
          description: Import committed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportBatch"
        "422":
          description: Import has blocking validation errors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /imports/validate:
    post:
      tags: [imports]
      operationId: validateImport
      summary: Validate an import file and preview rows without committing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportValidationInput"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportValidationResult"
  /imports/template:
    get:
      tags: [imports]
      operationId: getImportTemplate
      summary: Download a CSV template for an import type
      parameters:
        - name: importType
          in: query
          required: true
          schema:
            type: string
            enum: [VENDOR_MASTER, PO_REFERENCE, INVOICE_CORRECTION]
      responses:
        "200":
          description: CSV template
          content:
            text/csv:
              schema:
                type: string
                format: binary
  /imports/{id}:
    get:
      tags: [imports]
      operationId: getImport
      summary: Get an import batch by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Import batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportBatch"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ─── Phase 2: Exports (file-based; NOT ERP integration) ──────────────────────
  /exports:
    get:
      tags: [exports]
      operationId: listExports
      summary: List export batch history
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Export batches
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportBatchListResponse"
    post:
      tags: [exports]
      operationId: createExport
      summary: Generate an export file package and record the batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportRequest"
      responses:
        "201":
          description: Export batch created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportBatch"
  /exports/{id}:
    get:
      tags: [exports]
      operationId: getExport
      summary: Get an export batch by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Export batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportBatch"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /exports/{id}/download:
    get:
      tags: [exports]
      operationId: downloadExport
      summary: Download the CSV file for an export batch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: CSV file download
          content:
            text/csv:
              schema:
                type: string
                format: binary

  # ─── Phase 2: Accuracy reporting framework ───────────────────────────────────
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ─── Phase 2: Dashboard & Analytics ──────────────────────────────────────
...
    ImportRowError:
      type: object
      required: [row, message]
      properties:
        row:
          type: integer
        column:
          type: ["string", "null"]
        message:
          type: string
    ImportPreviewRow:
      type: object
      required: [rowNumber, valid, data, errors]
      properties:
        rowNumber:
          type: integer
        valid:
          type: boolean
        data:
          type: object
          additionalProperties:
            type: string
        errors:
          type: array
          items:
            type: string
    ImportValidationInput:
      type: object
      required: [importType, fileName, content]
      properties:
        importType:
          type: string
          enum: [VENDOR_MASTER, PO_REFERENCE, INVOICE_CORRECTION]
        fileName:
          type: string
          minLength: 1
        content:
          type: string
          description: Raw CSV file content.
    ImportValidationResult:
      type: object
      required:
        - importType
        - fileName
        - columns
        - rowCount
        - rowsValid
        - rowsRejected
        - preview
        - hasBlockingErrors
      properties:
        importType:
          type: string
        fileName:
          type: string
        columns:
          type: array
          items:
            type: string
        rowCount:
          type: integer
        rowsValid:
          type: integer
        rowsRejected:
          type: integer
        preview:
          type: array
          items:
            $ref: "#/components/schemas/ImportPreviewRow"
        errorSummary:
          type: ["string", "null"]
        hasBlockingErrors:
          type: boolean
    ImportCommitInput:
      type: object
      required: [importType, fileName, content]
      properties:
        importType:
          type: string
          enum: [VENDOR_MASTER, PO_REFERENCE, INVOICE_CORRECTION]
        fileName:
          type: string
          minLength: 1
        content:
          type: string
        uploadedBy:
          type: ["string", "null"]
        updateExisting:
          type: boolean
          default: false
          description: When true, existing rows (by natural key) are updated; otherwise duplicates are skipped.
    ImportBatch:
      type: object
      required:
        - id
        - batchId
        - importType
        - fileName
        - uploadedAt
        - rowCount
        - rowsAccepted
        - rowsRejected
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        batchId:
          type: string
        importType:
          type: string
        fileName:
          type: string
        uploadedBy:
          type: ["string", "null"]
        uploadedAt:
          type: string
          format: date-time
        rowCount:
          type: integer
        rowsAccepted:
          type: integer
        rowsRejected:
          type: integer
        status:
          type: string
          enum: [PENDING, VALIDATED, COMMITTED, CANCELLED, FAILED]
        errorSummary:
          type: ["string", "null"]
        rowErrors:
          type: array
          items:
            $ref: "#/components/schemas/ImportRowError"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ImportBatchListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ImportBatch"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── Phase 2: Exports ────────────────────────────────────────────────────
    ExportRequest:
      type: object
      required: [exportType]
      properties:
        exportType:
          type: string
          enum:
            - AP_INVOICE_FILE
            - APPROVED
            - POSTED
            - ALL_ACTIVE
            - EXCEPTIONS
            - TIE_OUT_FAILURES
            - VENDOR_SUMMARY
            - SOURCE_DOCUMENT_SUMMARY
        format:
          type: string
          enum: [CSV]
          default: CSV
        status:
          type: ["string", "null"]
        dateFrom:
          type: ["string", "null"]
        dateTo:
          type: ["string", "null"]
        vendorId:
          type: ["integer", "null"]
        exportedBy:
          type: ["string", "null"]
    ExportBatch:
      type: object
      required:
        - id
        - batchId
        - exportType
        - format
        - recordCount
        - exportedAt
        - status
        - createdAt
      properties:
        id:
          type: integer
        batchId:
          type: string
        exportType:
          type: string
        format:
          type: string
        filterJson:
          type: object
          additionalProperties: true
        recordCount:
          type: integer
        exportedBy:
          type: ["string", "null"]
        exportedAt:
          type: string
          format: date-time
        fileName:
          type: ["string", "null"]
        fileObjectPath:
          type: ["string", "null"]
        status:
          type: string
          enum: [SUCCESS, FAILED]
        createdAt:
          type: string
          format: date-time
    ExportBatchListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ExportBatch"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── Phase 2: Accuracy ───────────────────────────────────────────────────
...