# lib/api-zod/src/generated/types/exportBatch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Invoice Capture MVP API
 * OpenAPI spec version: 0.1.0
 */
import type { ExportBatchFilterJson } from './exportBatchFilterJson';
import type { ExportBatchStatus } from './exportBatchStatus';

export interface ExportBatch {
  id: number;
  batchId: string;
  exportType: string;
  format: string;
  filterJson?: ExportBatchFilterJson;
  recordCount: number;
  /** @nullable */
  exportedBy?: string | null;
  exportedAt: Date;
  /** @nullable */
  fileName?: string | null;
  /** @nullable */
  fileObjectPath?: string | null;
  status: ExportBatchStatus;
  createdAt: Date;
}