# lib/api-zod/src/generated/types/importRowError.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Invoice Capture MVP API
 * OpenAPI spec version: 0.1.0
 */

export interface ImportRowError {
  row: number;
  /** @nullable */
  column?: string | null;
  message: string;
}